refactor: move files with database to persistence directory
