+++ src/main/java/dev/tssvett/schedule_bot/backend/dto/StudentInfoDto.java
@@ -1,8 +1,8 @@
 package dev.tssvett.schedule_bot.backend.dto;
 
-import dev.tssvett.schedule_bot.backend.entity.Faculty;
-import dev.tssvett.schedule_bot.backend.entity.Group;
-import dev.tssvett.schedule_bot.backend.entity.Notification;
+import dev.tssvett.schedule_bot.persistence.entity.Faculty;
+import dev.tssvett.schedule_bot.persistence.entity.Group;
+import dev.tssvett.schedule_bot.persistence.entity.Notification;
 import dev.tssvett.schedule_bot.bot.enums.RegistrationState;
 import jakarta.validation.constraints.NotNull;
