+++ src/main/java/dev/tssvett/schedule_bot/backend/service/GroupService.java
@@ -1,8 +1,8 @@
 package dev.tssvett.schedule_bot.backend.service;
 
-import dev.tssvett.schedule_bot.backend.entity.Group;
+import dev.tssvett.schedule_bot.persistence.entity.Group;
 import dev.tssvett.schedule_bot.backend.exception.database.GroupNotExistException;
-import dev.tssvett.schedule_bot.backend.repository.GroupRepository;
+import dev.tssvett.schedule_bot.persistence.repository.GroupRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
