+++ src/main/java/dev/tssvett/schedule_bot/bot/keyboard/impl/course/CourseKeyboardButton.java
@@ -1,6 +1,6 @@
 package dev.tssvett.schedule_bot.bot.keyboard.impl.course;
 
-import dev.tssvett.schedule_bot.backend.entity.Student;
+import dev.tssvett.schedule_bot.persistence.entity.Student;
 import dev.tssvett.schedule_bot.backend.exception.registration.NotValidRegistrationStateException;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.actions.keyboard.impl.details.CallbackDetails;
