+++ src/main/java/dev/tssvett/schedule_bot/bot/keyboard/impl/faculty/FacultyKeyboard.java
@@ -1,6 +1,6 @@
 package dev.tssvett.schedule_bot.bot.keyboard.impl.faculty;
 
-import dev.tssvett.schedule_bot.backend.entity.Faculty;
+import dev.tssvett.schedule_bot.persistence.entity.Faculty;
 import dev.tssvett.schedule_bot.backend.service.FacultyService;
 import dev.tssvett.schedule_bot.bot.enums.Action;
 import dev.tssvett.schedule_bot.bot.keyboard.Keyboard;
