+++ src/main/java/dev/tssvett/schedule_bot/bot/keyboard/impl/notification/NotificationKeyboardButton.java
@@ -1,6 +1,6 @@
 package dev.tssvett.schedule_bot.bot.keyboard.impl.notification;
 
-import dev.tssvett.schedule_bot.backend.entity.Student;
+import dev.tssvett.schedule_bot.persistence.entity.Student;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.actions.keyboard.impl.details.CallbackDetails;
 import dev.tssvett.schedule_bot.bot.keyboard.KeyboardButton;
