+++ src/main/java/dev/tssvett/schedule_bot/parsing/SchoolWeekParser.java
@@ -1,7 +1,7 @@
 package dev.tssvett.schedule_bot.parsing;
 
 import dev.tssvett.schedule_bot.backend.exception.parse.ParserSourceConnectionException;
-import dev.tssvett.schedule_bot.backend.entity.Lesson;
+import dev.tssvett.schedule_bot.persistence.entity.Lesson;
 import lombok.extern.slf4j.Slf4j;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
