+++ src/main/java/dev/tssvett/schedule_bot/persistence/entity/Faculty.java
@@ -1,12 +1,10 @@
-package dev.tssvett.schedule_bot.backend.entity;
+package dev.tssvett.schedule_bot.persistence.entity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
@@ -18,27 +16,23 @@ import lombok.Setter;
 import java.util.Set;
 
 @Entity
-@Table(name = "\"group\"")
+@Table(name = "faculty")
 @Builder
 @Getter
 @Setter
 @RequiredArgsConstructor
 @AllArgsConstructor
-public class Group {
+public class Faculty {
     @Id
-    @Column(name = "group_id", nullable = false)
-    private Long groupId;
+    @Column(name = "faculty_id", nullable = false)
+    private Long facultyId;
 
     @Column(name = "name", nullable = false)
     private String name;
 
-    @Column(name = "course", nullable = false)
-    private Long course;
+    @OneToMany(mappedBy = "faculty", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    private Set<Group> groups;
 
-    @ManyToOne
-    @JoinColumn(name = "faculty_id", nullable = false)
-    private Faculty faculty;
-
-    @OneToMany(mappedBy = "group", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @OneToMany(mappedBy = "faculty", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private Set<Student> students;
 }
