+++ src/main/java/dev/tssvett/schedule_bot/persistence/entity/Group.java
@@ -1,10 +1,12 @@
-package dev.tssvett.schedule_bot.backend.entity;
+package dev.tssvett.schedule_bot.persistence.entity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
@@ -16,23 +18,27 @@ import lombok.Setter;
 import java.util.Set;
 
 @Entity
-@Table(name = "faculty")
+@Table(name = "\"group\"")
 @Builder
 @Getter
 @Setter
 @RequiredArgsConstructor
 @AllArgsConstructor
-public class Faculty {
+public class Group {
     @Id
-    @Column(name = "faculty_id", nullable = false)
-    private Long facultyId;
+    @Column(name = "group_id", nullable = false)
+    private Long groupId;
 
     @Column(name = "name", nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "faculty", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private Set<Group> groups;
+    @Column(name = "course", nullable = false)
+    private Long course;
 
-    @OneToMany(mappedBy = "faculty", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @ManyToOne
+    @JoinColumn(name = "faculty_id", nullable = false)
+    private Faculty faculty;
+
+    @OneToMany(mappedBy = "group", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private Set<Student> students;
 }
