+++ src/main/java/dev/tssvett/schedule_bot/scheduling/SchedulingNotification.java
@@ -1,8 +1,8 @@
 package dev.tssvett.schedule_bot.scheduling;
 
-import dev.tssvett.schedule_bot.backend.entity.Student;
-import dev.tssvett.schedule_bot.backend.entity.Lesson;
-import dev.tssvett.schedule_bot.backend.entity.Notification;
+import dev.tssvett.schedule_bot.persistence.entity.Student;
+import dev.tssvett.schedule_bot.persistence.entity.Lesson;
+import dev.tssvett.schedule_bot.persistence.entity.Notification;
 import dev.tssvett.schedule_bot.backend.service.NotificationService;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.TelegramBot;
